R120

+++ cryptohome/storage/encrypted_container/loopback_device.cc
@@ -1,39 +1,37 @@
-// Copyright 2020 The Chromium OS Authors. All rights reserved.
+// Copyright 2020 The ChromiumOS Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 #include "cryptohome/storage/encrypted_container/loopback_device.h"
 
 #include <memory>
+#include <optional>
 #include <string>
 #include <utility>
 
 #include <base/files/file_path.h>
 #include <base/logging.h>
 #include <base/values.h>
-#include <brillo/blkdev_utils/loop_device.h>
 
 #include "cryptohome/platform.h"
 #include "cryptohome/storage/encrypted_container/backing_device.h"
 
 namespace cryptohome {
 
-LoopbackDevice::LoopbackDevice(
-    const BackingDeviceConfig& config,
-    Platform* platform,
-    std::unique_ptr<brillo::LoopDeviceManager> loop_device_manager)
-    : name_(config.name),
-      size_(config.size),
-      backing_file_path_(config.loopback.backing_file_path),
-      platform_(platform),
-      loop_device_manager_(std::move(loop_device_manager)) {}
-
 LoopbackDevice::LoopbackDevice(const BackingDeviceConfig& config,
                                Platform* platform)
-    : LoopbackDevice(
-          config, platform, std::make_unique<brillo::LoopDeviceManager>()) {}
+    : backing_file_path_(config.loopback.backing_file_path),
+      name_(config.name),
+      size_(config.size),
+      fixed_backing_(config.loopback.fixed_backing),
+      platform_(platform) {}
 
 bool LoopbackDevice::Create() {
+  if (fixed_backing_) {
+    LOG(ERROR) << "Backing of loop device " << backing_file_path_
+               << " cannot be created, as it is fixed.";
+    return false;
+  }
   if (!platform_->CreateSparseFile(backing_file_path_, size_) ||
       !platform_->SetPermissions(backing_file_path_, S_IRUSR | S_IWUSR)) {
     LOG(ERROR) << "Failed to create sparse file.";
@@ -43,13 +41,20 @@ bool LoopbackDevice::Create() {
 }
 
 bool LoopbackDevice::Purge() {
+  if (fixed_backing_) {
+    LOG(ERROR) << "Fixed backing storage cannot be purged: "
+               << backing_file_path_;
+
+    return false;
+  }
+
   return platform_->DeleteFile(backing_file_path_);
 }
 
 bool LoopbackDevice::Setup() {
   // Set up loopback device.
   std::unique_ptr<brillo::LoopDevice> loopdev =
-      loop_device_manager_->AttachDeviceToFile(backing_file_path_);
+      platform_->GetLoopDeviceManager()->AttachDeviceToFile(backing_file_path_);
 
   if (!loopdev->IsValid()) {
     LOG(ERROR) << "Failed to attach loop back device";
@@ -68,12 +73,14 @@ bool LoopbackDevice::Setup() {
 
 bool LoopbackDevice::Teardown() {
   std::unique_ptr<brillo::LoopDevice> loopdev =
-      loop_device_manager_->GetAttachedDeviceByName(name_);
+      platform_->GetLoopDeviceManager()->GetAttachedDeviceByName(name_);
 
   if (!loopdev->IsValid()) {
     LOG(ERROR) << "Loop device does not exist.";
     return false;
   }
+
+  std::ignore = loopdev->SetName("");
 
   return loopdev->Detach();
 }
@@ -82,13 +89,13 @@ bool LoopbackDevice::Exists() {
   return platform_->FileExists(backing_file_path_);
 }
 
-base::Optional<base::FilePath> LoopbackDevice::GetPath() {
+std::optional<base::FilePath> LoopbackDevice::GetPath() {
   std::unique_ptr<brillo::LoopDevice> loopdev =
-      loop_device_manager_->GetAttachedDeviceByName(name_);
+      platform_->GetLoopDeviceManager()->GetAttachedDeviceByName(name_);
 
   if (!loopdev->IsValid()) {
     LOG(ERROR) << "Loop device does not exist.";
-    return base::nullopt;
+    return std::nullopt;
   }
 
   return loopdev->GetDevicePath();
